Subscription dialog confirm

+++ front-end/src/app/components/subscription/subscription.component.ts
@@ -3,6 +3,8 @@ import { Router } from '@angular/router';
 import { AuthService } from '../auth.service';
 import { SubscriptionService } from '../subscription.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogModel, ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-subscription',
@@ -11,10 +13,12 @@ import { MatSnackBar } from '@angular/material/snack-bar';
 })
 export class SubscriptionComponent implements OnInit {
 
-  constructor(private snackbar: MatSnackBar, private router: Router, public authService: AuthService, public subscriptionService: SubscriptionService) { }
+  result: string = '';
+
+  constructor(public dialog: MatDialog, private snackbar: MatSnackBar, private router: Router, public authService: AuthService, public subscriptionService: SubscriptionService) { }
 
   ngOnInit(): void {
-    
+
   }
 
   hasSubscription(): boolean {
@@ -22,12 +26,25 @@ export class SubscriptionComponent implements OnInit {
   }
 
   deleteSubscription() {
-    let subscription = this.authService.getSubscription();
-    this.authService.setSubscription();
-    this.subscriptionService.deleteSubscription(subscription);
-    this.snackbar.open('Suscripción cancelada correctamente ', '', {
-      duration: 4000
+    const message = '¿Está seguro que desea cancelar su suscripción?';
+    const dialogData = new ConfirmDialogModel("Cancelar suscripción", message);
+
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      maxWidth: '400px',
+      data: dialogData
     });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        let subscription = this.authService.getSubscription();
+        this.authService.setSubscription();
+        this.subscriptionService.deleteSubscription(subscription);
+        this.snackbar.open('Suscripción cancelada correctamente ', '', {
+          duration: 4000
+        });
+      }
+    });
+
   }
 
   isLoggedIn(): boolean {
